TPCircularBuffer update to avoid a 64-bit warning

+++ TheAmazingAudioEngine/Library/TPCircularBuffer/TPCircularBuffer+AudioBufferList.c
@@ -197,9 +197,9 @@ void TPCircularBufferConsumeNextBufferListPartial(TPCircularBuffer *buffer, int 
     // Reposition block forward, just before the audio data, ensuring 16-byte alignment
     TPCircularBufferABLBlockHeader *newBlock = (TPCircularBufferABLBlockHeader*)(((unsigned long)block + bytesToConsume) & ~0xFul);
     memmove(newBlock, block, sizeof(TPCircularBufferABLBlockHeader) + (block->bufferList.mNumberBuffers-1)*sizeof(AudioBuffer));
-    int32_t bytesFreed = (int32_t)newBlock - (int32_t)block;
+    intptr_t bytesFreed = (intptr_t)newBlock - (intptr_t)block;
     newBlock->totalLength -= bytesFreed;
-    TPCircularBufferConsume(buffer, bytesFreed);
+    TPCircularBufferConsume(buffer, (int32_t)bytesFreed);
 }
 
 void TPCircularBufferDequeueBufferListFrames(TPCircularBuffer *buffer, UInt32 *ioLengthInFrames, AudioBufferList *outputBufferList, AudioTimeStamp *outTimestamp, const AudioStreamBasicDescription *audioFormat) {
